+++ codemods/debarrel/tests/package-name-alias-namespace-barrel/expected/src/z-widget.stories.tsx
@@ -1,0 +1,3 @@
+import * as Storybook from "sentry/stories";
+
+console.log(Storybook.story());